ChatGPT rewrite
